backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.2.7
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-autoslug==1.9.9
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.13.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
kombu==5.3.1
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg==3.1.9
psycopg-binary==3.1.9
pycparser==2.21
PyJWT==2.7.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
